fix: 修复 sys_print_task 与 sys_print_chain_item 表 JOIN 类型不匹配
- chain_item_id(bigint) 与 item_key(varchar) 比较需要显式类型转换

+++ backend/erp/erp-printing/src/main/java/cn/aiedge/erp/printing/mapper/SysScreenshotTaskMapper.java
@@ -20,7 +20,7 @@ public interface SysScreenshotTaskMapper extends BaseMapper<SysScreenshotTask> {
     @Select("""
             SELECT ss.* FROM sys_screenshot_task ss
             JOIN sys_print_task t ON ss.screenshot_id = t.screenshot_id
-            JOIN sys_print_chain_item ci ON t.chain_item_id = ci.item_key
+            JOIN sys_print_chain_item ci ON t.chain_item_id = ci.item_key::bigint
             WHERE ss.status = 'PENDING'
               AND ci.screenshot_mode = 'MANUAL_CONFIRM'
               AND ci.screenshot_confirm_timeout IS NOT NULL
